re-order project layout to make it more managable
