+++ src/app/app.routes.ts
@@ -1,14 +1,14 @@
 import { Routes } from '@angular/router';
-import { ImprintComponent } from './imprint/imprint.component';
-import { StatusViewComponent } from './status-view/status-view.component';
-import { CookiesComponent } from './cookies/cookies.component';
-import { IncidentDetailsViewComponent } from './incident-details-view/incident-details-view.component';
-import { PageNotFoundComponent } from './page-not-found/page-not-found.component';
-import { ManagementViewComponent } from './management-view/management-view.component';
+import { ImprintComponent } from './pages/imprint/imprint.component';
+import { StatusViewComponent } from './pages/status-view/status-view.component';
+import { CookiesComponent } from './pages/cookies/cookies.component';
+import { IncidentDetailsViewComponent } from './pages/incident-details-view/incident-details-view.component';
+import { PageNotFoundComponent } from './pages/page-not-found/page-not-found.component';
+import { ManagementViewComponent } from './pages/management-view/management-view.component';
 import { autoLoginPartialRoutesGuard } from 'angular-auth-oidc-client';
-import { ForbiddenComponent } from './forbidden/forbidden.component';
-import { UnauthorizedComponent } from './unauthorized/unauthorized.component';
-import { LoginComponent } from './login/login.component';
+import { ForbiddenComponent } from './pages/forbidden/forbidden.component';
+import { UnauthorizedComponent } from './pages/unauthorized/unauthorized.component';
+import { LoginComponent } from './pages/login/login.component';
 
 export const routes: Routes = [
     {path: "", component: StatusViewComponent},
