+++ src/app/component-list/component-list.component.ts
@@ -1,8 +1,8 @@
 import { Component as AComponent, Input } from '@angular/core';
 import { Component } from '../../external/lib/status-page-api/angular-client'; 
 import { CommonModule } from '@angular/common';
-import { AppConfigService } from '../app-config.service';
-import { UtilService } from '../util.service';
+import { AppConfigService } from '../services/app-config.service';
+import { UtilService } from '../services/util.service';
 import { DailyStatus } from '../model/daily-status';
 import dayjs from 'dayjs';
 import { ComponentId } from '../model/base';
