+++ src/app/footer/footer.component.ts
@@ -1,6 +1,6 @@
 import { CommonModule } from '@angular/common';
 import { Component, inject } from '@angular/core';
-import { UserSettingsService } from '../user-settings.service';
+import { UserSettingsService } from '../services/user-settings.service';
 import { FormsModule } from '@angular/forms';
 import { RouterModule } from '@angular/router';
 import { OidcSecurityService } from 'angular-auth-oidc-client';
