+++ src/app/incident-log-list/incident-log-list.component.ts
@@ -3,7 +3,7 @@ import { Component, Input } from '@angular/core';
 import { Incident } from '../../external/lib/status-page-api/angular-client'; 
 import { ReversePipe } from '../reverse.pipe';
 import { RouterModule } from '@angular/router';
-import { UtilService } from '../util.service';
+import { UtilService } from '../services/util.service';
 import { IncidentId } from '../model/base';
 
 @Component({
