+++ src/app/pages/cookies/cookies.component.html
