+++ src/app/pages/cookies/cookies.component.spec.ts
