+++ src/app/pages/cookies/cookies.component.ts
