+++ src/app/pages/incident-details-view/incident-details-view.component.html
