+++ src/app/pages/management-view/management-view.component.css
