+++ src/app/pages/management-view/management-view.component.html
