+++ src/app/pages/management-view/management-view.component.ts
@@ -1,17 +1,17 @@
 import { CommonModule } from '@angular/common';
 import { Component, ElementRef, OnInit, Signal, ViewChild } from '@angular/core';
 import { Router, RouterModule } from '@angular/router';
-import { DataService } from '../data.service';
+import { DataService } from '../../services/data.service';
 import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
-import { UtilService } from '../util.service';
-import { Incident, IncidentService, IncidentUpdate, IncidentUpdateResponseData } from '../../external/lib/status-page-api/angular-client';
+import { UtilService } from '../../services/util.service';
+import { Incident, IncidentService, IncidentUpdate, IncidentUpdateResponseData } from '../../../external/lib/status-page-api/angular-client';
 import dayjs from 'dayjs';
-import { formatQueryDate, IncidentId } from '../model/base';
-import { SpinnerComponent } from '../spinner/spinner.component';
+import { formatQueryDate, IncidentId } from '../../model/base';
+import { SpinnerComponent } from '../../spinner/spinner.component';
 import { OidcSecurityService, UserDataResult } from 'angular-auth-oidc-client';
 import { firstValueFrom, Observable } from 'rxjs';
-import { AppConfigService } from '../app-config.service';
-import { IconProviderService } from '../icon-provider.service';
+import { AppConfigService } from '../../services/app-config.service';
+import { IconProviderService } from '../../services/icon-provider.service';
 
 const DT_FORMAT = "YYYY-MM-DDTHH:mm";
