+++ src/app/pages/status-view/status-view.component.spec.ts
