+++ src/app/services/data.service.ts
@@ -3,9 +3,9 @@ import dayjs from 'dayjs';
 import { HttpClient } from '@angular/common/http';
 import { AppConfigService } from './app-config.service';
 import { BehaviorSubject, Observable, combineLatestWith } from 'rxjs';
-import { DailyStatus } from './model/daily-status';
-import { Component, ComponentService, IdField, ImpactService, ImpactType, Incident, IncidentResponseData, IncidentService, IncidentUpdateResponseData, PhaseList, PhaseService, Severity } from '../external/lib/status-page-api/angular-client';
-import { ComponentId, formatQueryDate, ImpactId, IncidentId, SHORT_DAY_FORMAT, ShortDayString } from './model/base';
+import { DailyStatus } from '../model/daily-status';
+import { Component, ComponentService, IdField, ImpactService, ImpactType, Incident, IncidentResponseData, IncidentService, IncidentUpdateResponseData, PhaseList, PhaseService, Severity } from '../../external/lib/status-page-api/angular-client';
+import { ComponentId, formatQueryDate, ImpactId, IncidentId, SHORT_DAY_FORMAT, ShortDayString } from '../model/base';
 
 @Injectable({
   providedIn: 'root'
