+++ src/app/services/icon-provider.service.spec.ts
