+++ src/app/services/icon-provider.service.ts
